+++ server/core/app.go
@@ -6,11 +6,9 @@ import (
 	"context"
 	"fmt"
 	"log/slog"
-	"strings"
 	"time"
 
 	"github.com/jmoiron/sqlx"
-	"github.com/marcboeker/go-duckdb/v2"
 	"github.com/nats-io/nats.go"
 	"github.com/nats-io/nats.go/jetstream"
 )
@@ -311,120 +309,6 @@ func initDuckDB(duckDbx *sqlx.DB) error {
 		if err := createType(duckDbx, t.Name, t.Definition); err != nil {
 			return fmt.Errorf("failed to create custom type %s: %w", t.Name, err)
 		}
-	}
-	return nil
-}
-
-func migrateSystemData(sqliteDbx *sqlx.DB, duckDbx *sqlx.DB, deprecatedSchema string, logger *slog.Logger) error {
-	for _, table := range []string{
-		"apps",
-		"api_keys",
-		"users",
-		"sessions",
-		"invites",
-		"task_runs",
-	} {
-		if err := migrateTableData(sqliteDbx, duckDbx, table, deprecatedSchema, logger); err != nil {
-			return fmt.Errorf("failed to migrate table %s: %w", table, err)
-		}
-	}
-	return nil
-}
-
-// The database contains little data so we can read all data into memory and write it to sqlite database.
-// We don't delete data from duckdb after migrating.
-// We only migrate data if the target table is empty.
-func migrateTableData(sqliteDbx *sqlx.DB, duckDbx *sqlx.DB, table string, deprecatedSchema string, logger *slog.Logger) error {
-	duckDbTable := "\"" + deprecatedSchema + "\"." + table
-	var duckCount int
-	err := duckDbx.Get(&duckCount, fmt.Sprintf(`SELECT count(*) FROM %s`, duckDbTable))
-	// Skip if table not in DuckDB
-	if err != nil {
-		return nil
-	}
-	if duckCount == 0 {
-		// DuckDB table has no data, skipping
-		return nil
-	}
-	// Check if table already has data
-	var liteCount int
-	err = sqliteDbx.Get(&liteCount, fmt.Sprintf(`SELECT count(*) FROM %s`, table))
-	if err != nil {
-		return fmt.Errorf("failed to count rows in sqlite table %s: %w", table, err)
-	}
-	if liteCount > 0 {
-		// Table already has data, skip migration
-		return nil
-	}
-	logger.Info("Migrating table to SQLite", slog.String("table", table), slog.Int("rows", duckCount))
-	rows, err := duckDbx.Queryx(fmt.Sprintf(`SELECT * FROM %s`, duckDbTable))
-	if err != nil {
-		return fmt.Errorf("failed to query duckdb table %s: %w", duckDbTable, err)
-	}
-	defer rows.Close()
-	cols, err := rows.Columns()
-	if err != nil {
-		return fmt.Errorf("failed to get columns for duckdb table %s: %w", duckDbTable, err)
-	}
-	tx, err := sqliteDbx.Begin()
-	if err != nil {
-		return fmt.Errorf("failed to begin transaction for sqlite table %s: %w", table, err)
-	}
-	// Migrate each row
-	for rows.Next() {
-		values := make([]any, len(cols))
-		valuePtrs := make([]any, len(cols))
-		for i := range cols {
-			valuePtrs[i] = &values[i]
-		}
-		if err := rows.Scan(valuePtrs...); err != nil {
-			tx.Rollback()
-			return fmt.Errorf("failed to scan row for duckdb table %s: %w", duckDbTable, err)
-		}
-		placeholders := ""
-		for i := range cols {
-			if i > 0 {
-				placeholders += ", "
-			}
-			placeholders += "?"
-			if values[i] != nil {
-				// if v, ok := values[i].(time.Time); ok {
-				// 	values[i] = v.UnixMilli()
-				// }
-				if table == "task_runs" {
-					if cols[i] == "last_run_duration" {
-						v, ok := values[i].(duckdb.Interval)
-						if !ok {
-							tx.Rollback()
-							return fmt.Errorf("failed to convert last_run_duration to time for duckdb table %s", table)
-						}
-						values[i] = formatInterval(v)
-						// } else if cols[i] == "last_run_success" {
-						// 	v, ok := values[i].(bool)
-						// 	if !ok {
-						// 		tx.Rollback()
-						// 		return fmt.Errorf("failed to convert last_run_success to bool for duckdb table %s", table)
-						// 	}
-						// 	if v {
-						// 		values[i] = 1
-						// 	} else {
-						// 		values[i] = 0
-						// 	}
-					}
-				}
-			}
-		}
-		_, err := tx.Exec(
-			fmt.Sprintf(`INSERT INTO %s (%s) VALUES (%s)`, table, strings.Join(cols, ", "), placeholders),
-			values...,
-		)
-		if err != nil {
-			tx.Rollback()
-			return fmt.Errorf("failed to insert row into sqlite table %s: %w", table, err)
-		}
-	}
-	if err := tx.Commit(); err != nil {
-		return fmt.Errorf("failed to commit transaction for sqlite table %s: %w", table, err)
 	}
 	return nil
 }
